fixed all the breaking errors and everything is like how it was before and also if you swipe all of it goes to the results

+++ app/src/main/java/com/example/ibuprofen/Adapters/RestaurantsAdapter.java
@@ -97,7 +97,7 @@ public class RestaurantsAdapter extends RecyclerView.Adapter<RestaurantsAdapter.
             tvName.setText(restaurant.getName());
             tvCuisine.setText(restaurant.getCategories());
             tvDistance.setText(String.format("%.2f miles", restaurant.getDistance()));
-            rbRating.setRating(restaurant.getRating());
+            rbRating.setRating((float) restaurant.getRating());
             tvPrice.setText(restaurant.getPrice());
         }
     }
